elf: Move relocation handling code to separate module
Move all structs and code handling ELF relocations to a new submodule.
